feat: delete user fetch login in its comonente
